AA-package-driver/requirements.txt
ollama
openai-whisper
pyaudio

websocket-client
websockets
requests
fastapi
uvicorn
pydantic
